Preserve double precision in FIX

+++ Trs80.Level1Basic.VirtualMachine/Machine/Trs80Api.cs
@@ -106,7 +106,7 @@ public class Trs80Api : ITrs80Api
 
     public int Fix(dynamic value)
     {
-        float numericValue = Convert.ToSingle(value, System.Globalization.CultureInfo.InvariantCulture);
+        double numericValue = Convert.ToDouble(value, System.Globalization.CultureInfo.InvariantCulture);
         return (int)Math.Truncate(numericValue);
     }
